[refactor] user api명을 auth로 변경
- 권한 관련 api들이라 user에서 auth로 이름 변경
- 서버와 맞춰주었다
